QtQml: Explicitly byte-swap on big endian when building strings from CU
Commit 9ef4c123c39c642357c9e8530d59f32f220a7824 in qtbase made the
implicit conversions to QChar through one of the integer types invalid.
quint16_le can cast to quint16 (at a cost) but isn't quint16. Doing an
explicit casting is required now.

+++ src/qml/common/qv4compileddata_p.h
@@ -1297,7 +1297,7 @@ struct Unit
         QString qstr(str->size, Qt::Uninitialized);
         QChar *ch = qstr.data();
         for (int i = 0; i < str->size; ++i)
-             ch[i] = QChar(characters[i]);
+             ch[i] = QChar(quint16(characters[i]));
          return qstr;
 #endif
     }
